server_thread_id_connect

+++ 함범호/sever_test/LabProject07-9-7(Animation) - 복사본/Server.h
@@ -26,7 +26,7 @@ struct Player_world {
 #pragma pack(push,1)
 struct Thread_id {
 
-	int thread_id;
+	short thread_id;
 
 };
 #pragma pack(pop)
