+++ 함범호/sever_test/Sever_test/Sever_test/framework.h
@@ -71,7 +71,7 @@ struct Player_world {
 #pragma pack(push,1)
 struct Thread_id {
 
-	int thread_id = 0;
+	short thread_id = 0;
 
 };
 #pragma pack(pop)
